platform_backend/pyproject.toml
license = {text = "MIT"}
readme = "README.md"

[tool.poetry]
name = "HR_Platform"
version = "0.1.0"
description = ""
authors = ["Amirhan Botagariev <ruby.mainacc@gmail.com>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.8"
uvicorn = {extras = ["standard"], version = "^0.34.0"}
pydantic = {extras = ["email"], version = "^2.10.6"}
pydantic-settings = "^2.7.1"
black = "^25.1.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.37"}
asyncpg = "^0.30.0"
alembic = "^1.14.1"
orjson = "^3.10.15"
fastapi-users = {extras = ["sqlaclhemy"], version = "^14.0.1"}
fastapi-users-db-sqlalchemy = "^7.0.0"
gunicorn = "^23.0.0"
joblib = "^1.4.2"
pandas = "^2.2.3"
scikit-learn = "^1.6.1"
pyjwt = "^2.10.1"
sqlparse = "^0.5.3"
requests = "^2.32.3"

[tool.black]
line-length = 88
target-version = ["py312"]
